requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.32.0
azure-identity==1.21.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
msal==1.32.0
msal-extensions==1.3.1
mypy-extensions==1.0.0
openai==1.68.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyinvoke==1.0.4
PyJWT==2.10.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruff==0.11.2
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
